Protection api

+++ server/api/modifiable/deleteActu.post.ts
@@ -1,8 +1,19 @@
+
+import { getServerSession } from '#auth'
 export default defineEventHandler(async (event) => {
     const body = await readBody(event)
 
-    await event.context.prisma.article.delete({
-        where: { id: body.id },
-    })
-    return { message: "User updated" }
+
+    const session = await getServerSession(event)
+    if (!session) {
+        return { message: "You are not authorized to perform this action" }
+    }
+    else{    
+        if ( session?.user?.id === body.id){
+            return { message: "User not created" }
+        } else {
+            await event.context.prisma.article.delete({
+                where: { id: body.id },
+            })
+    }}
 })
